:wrench: Add settings_user.yml v2
The original user settings yml file was far too complex. The creator of
conan saw the tree like structure for each architecture and stated that
the set of potential options would cause an explosion of binaries,
which is correct. This number can be brought down by using the cpu names
we commonly use with arm cortex devices.

V2 also adds the libc setting for the GCC compiler, allowing a selection
of .specs such as nano, nosys, and importantly picolibc.