:wrench: Move to LLVM & clang for all host platforms
The move to LLVM comes from the ease of installing it on older Ubuntu
systems (20.04 and below) using the llvm.sh script. Installation on
Mac and Windows is also quite easy given the use of homebrew and
chocolatey. LLVM also comes with everything we need for your toolchain
to be successful, specifically, clang-tidy, and clangd.